backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
django-filter==25.1
reportlab==4.2.5
google-auth==2.38.0
requests==2.32.3
python-dotenv==1.0.0

